pkg/bb: move bb creator package to pkg/bb
pkg/bb -> pkg/bb/bbmain
pkg/uroot/builder/bb -> pkg/bb
