+++ pkg/bb/bbmain/register.go
@@ -2,7 +2,9 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package bb
+// Package bbmain is the package imported by all rewritten busybox
+// command-packages to register themselves.
+package bbmain
 
 import (
 	"errors"
